requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
aiofiles==23.2.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
requests==2.32.3
scikit-learn==1.5.1
tensorflow==2.17.1
python-multipart==0.0.9
openpyxl==3.1.5
mangum==0.18.0
